# internal/db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlc

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Account struct {
	ID                string
	UserID            string
	Provider          string
	ProviderAccountID string
	Scopes            []string
	AccessToken       []byte
	AccessExpiresAt   pgtype.Timestamptz
	RefreshToken      []byte
	RefreshExpiresAt  pgtype.Timestamptz
	CreatedAt         pgtype.Timestamptz
	UpdatedAt         pgtype.Timestamptz
}

type ApiKey struct {
	ID         string
	UserID     string
	Name       string
	Prefix     string
	KeyHash    []byte
	LastUsedAt pgtype.Timestamptz
	CreatedAt  pgtype.Timestamptz
}

type Review struct {
	ID             string
	UserID         string
	Source         []byte
	Note           string
	Status         string
	Result         string
	Reason         string
	Attempts       int32
	IdempotencyKey *string
	RequestHash    []byte
	LeaseOwner     string
	LeaseToken     int64
	LeaseUntil     pgtype.Timestamptz
	CreatedAt      pgtype.Timestamptz
	StartedAt      pgtype.Timestamptz
	FinishedAt     pgtype.Timestamptz
}

type Session struct {
	ID        string
	UserID    string
	TokenHash []byte
	ExpiresAt pgtype.Timestamptz
	Ip        string
	UserAgent string
	CreatedAt pgtype.Timestamptz
}

type User struct {
	ID        string
	Email     string
	Name      string
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
}