# spec/swagger.yaml
components:
  schemas:
...
    CreateReviewRequest:
      properties:
        note:
          example: focus on auth
          type: string
        source:
          $ref: '#/components/schemas/Source'
      required:
      - source
      type: object
...
    ErrorCode:
      enum:
      - bad_request
      - unauthenticated
      - not_found
      - method_not_allowed
      - internal
      - unavailable
      - review_not_found
      - idempotency_conflict
      - rate_limited
      - invalid_credentials
      - email_taken
      - invalid_email
      - weak_password
      - api_key_not_found
      - session_not_found
      - review_running
      - origin_forbidden
      - repository_not_found
      - provider_not_linked
      - provider_expired
      type: string
      x-enum-varnames:
      - CodeBadRequest
      - CodeUnauthenticated
      - CodeNotFound
      - CodeMethodNotAllowed
      - CodeInternal
      - CodeUnavailable
      - CodeReviewNotFound
      - CodeIdempotencyConflict
      - CodeRateLimited
      - CodeInvalidCredentials
      - CodeEmailTaken
      - CodeInvalidEmail
      - CodeWeakPassword
      - CodeAPIKeyNotFound
      - CodeSessionNotFound
      - CodeReviewRunning
      - CodeOriginForbidden
      - CodeRepositoryNotFound
      - CodeProviderNotLinked
      - CodeProviderExpired
...
    Problem:
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
          example: review_not_found
        detail:
          example: review r_123 does not exist
          type: string
        request_id:
          type: string
        status:
          example: 404
          type: integer
        title:
          example: Not Found
          type: string
        type:
          example: about:blank
          type: string
      required:
      - code
      - status
      - title
      - type
      type: object
...
    Review:
      properties:
        attempts:
          type: integer
        created_at:
          type: string
        finished_at:
          type: string
        id:
          example: r_9f2c1a7b3e5d4c60
          type: string
        note:
          type: string
        reason:
          example: the review pipeline is not implemented yet
          type: string
        result:
          enum:
          - complete
          - none
          type: string
        source:
          $ref: '#/components/schemas/Source'
        started_at:
          type: string
        status:
          enum:
          - queued
          - running
          - done
          type: string
      required:
      - attempts
      - created_at
      - id
      - source
      - status
      type: object
    ReviewList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Review'
          type: array
          uniqueItems: false
      required:
      - items
      type: object
...
    Source:
      properties:
        base:
          example: main
          type: string
        head:
          example: feature/login
          type: string
        repo:
          example: octo/app
          type: string
      required:
      - base
      - head
      - repo
      type: object
...
    UpdateUserRequest:
      properties:
        name:
          example: Ada Lovelace
          type: string
      required:
      - name
      type: object
    User:
      properties:
        created_at:
          type: string
        email:
          type: string
        id:
          example: usr_k3x9p2m4q8r1t6v0
          type: string
        name:
          type: string
      required:
      - created_at
      - email
      - id
      type: object
  securitySchemes:
    BearerAuth:
      description: Browser session set by sign-in
      in: cookie
      name: eyeful_session
      type: apiKey
...
info:
  description: Multi-agent code review with executable evidence.
  title: eyeful API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /reviews:
    get:
      operationId: listReviews
      parameters:
      - description: At most this many (1-100)
        in: query
        name: limit
        schema:
          default: 20
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewList'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
      security:
      - SessionCookie: []
      - BearerAuth: []
      summary: List reviews, newest first
      tags:
      - reviews
    post:
      description: Returns 202 with the queued review. A review cannot be stopped
        once it runs, so clients confirm before creating one. Repeating a request
        with the same Idempotency-Key returns the first review; reusing the key with
        a different body is a 409.
      operationId: createReview
      parameters:
      - description: Makes retries safe
        in: header
        name: Idempotency-Key
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
              description: What to review
              summary: body
        description: What to review
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Too Many Requests
      security:
      - SessionCookie: []
      - BearerAuth: []
      summary: Queue a review
      tags:
      - reviews
  /reviews/{id}:
    delete:
      operationId: deleteReview
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict
      security:
      - SessionCookie: []
      - BearerAuth: []
      summary: Delete a queued or done review
      tags:
      - reviews
    get:
      operationId: getReview
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
      security:
      - SessionCookie: []
      - BearerAuth: []
      summary: Get one review
      tags:
      - reviews
...
  /users/me:
    delete:
      operationId: deleteCurrentUser
      responses:
        "204":
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
      security:
      - SessionCookie: []
      - BearerAuth: []
      summary: Delete the signed-in user with their sessions, keys and reviews
      tags:
      - users
    get:
      operationId: getCurrentUser
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
      security:
      - SessionCookie: []
      - BearerAuth: []
      summary: The signed-in user
      tags:
      - users
    patch:
      operationId: updateCurrentUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
              description: New name
              summary: body
        description: New name
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
      security:
      - SessionCookie: []
      - BearerAuth: []
      summary: Change the signed-in user's name
      tags:
      - users
servers:
- url: /